back-end/requirements.txt
tensorflow
opencv-python
numpy
faiss-cpu
Pillow
scikit-learn
scikit-image 